Extract useRoomState hook into hooks directory
